requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
